refactor: move worker logic to separate file and improve message processing
